requirements.txt
discord.py==1.7.3
Jinja2==3.0.2
paramiko==2.11.0
requests==2.24.0
